Replacing unneeded test with more useful test
